Migrate Theme attributes to WithTheme
