+++ src/Traits/Helpers/ComponentHelpers.php
@@ -57,44 +57,6 @@ trait ComponentHelpers
         return $this->model;
     }
 
-    public function setTheme(): void
-    {
-        $theme = $this->getTheme();
-
-        if ($theme === 'bootstrap-4' || $theme === 'bootstrap-5') {
-            $this->setPaginationTheme('bootstrap');
-        }
-    }
-
-    public function getTheme(): string
-    {
-        return $this->theme ?? config('livewire-tables.theme', 'tailwind');
-    }
-
-    #[Computed]
-    public function isTailwind(): bool
-    {
-        return $this->getTheme() === 'tailwind';
-    }
-
-    #[Computed]
-    public function isBootstrap(): bool
-    {
-        return $this->getTheme() === 'bootstrap-4' || $this->getTheme() === 'bootstrap-5';
-    }
-
-    #[Computed]
-    public function isBootstrap4(): bool
-    {
-        return $this->getTheme() === 'bootstrap-4';
-    }
-
-    #[Computed]
-    public function isBootstrap5(): bool
-    {
-        return $this->getTheme() === 'bootstrap-5';
-    }
-
     /**
      * Get the translated empty message of the table
      */
